states and transitions for competence model (#31)
* Write state generator

* run black

* write & test types in state function

* write tests for potential states

* generate potential states

* adds test for potential states

* write tests for get_rate

* add retirement date to individual

tests are currently failing

* fix tests to pass with individuals retirement date

* move get_types_in_state_from_competence_state to states

* write functions to simulate retirement

For the competence model we will no longer
calculate the possible states and then getting
the transitios probabilities. Instead we will simulate
the system.

If the system is full one muast retire. The functions
for simulating a retirement have been implemented and tested

* write functions to simulate competence model

More specifically to simulate the hiring and
promoting processes. Tests for both have been
included

* rename functions provoke -> make

* minor tweaks

* fix retirement

We were retiring the person with the maximum retirement date instead of the
min. The last_retirement is renamed to last_retirement_date