write test for promotion that fails
Currently the promotion rate of an individual at level i is based
on the homophily effect of the people at level i. This however means
that the type of the top individual will never have an effect in any
promotion.

So it has been agreed that the promotion of an individual at level i
is affected by the types of individuals at level i + 1.

This test fails with the current implementation