Base model tweaks and implementation of competence model (#34)
* Add Jupyter to venv.

* Add matplotlib to venv

* Fix state generation.

* Fix transitions and improve simulation.

* Add a jupyter notebook

* Fix transition rates from going negative

* add bound case for capacity one

* force summit to be [1, 0]

* add example to compare theoretical and simulated

* Simulate stat vector (#21)

* implement fuction to get simulated stat vector

* run black

* first write up of model with competence (#26)

* fix test failures

* restructure code base

closes #23

* push files for restructure

* Write class for individual (#29)

* Write failing test for individual.

* Write init method for Individual class.

* states and transitions for competence model (#31)

* Write state generator

* run black

* write & test types in state function

* write tests for potential states

* generate potential states

* adds test for potential states

* write tests for get_rate

* add retirement date to individual

tests are currently failing

* fix tests to pass with individuals retirement date

* move get_types_in_state_from_competence_state to states

* write functions to simulate retirement

For the competence model we will no longer
calculate the possible states and then getting
the transitios probabilities. Instead we will simulate
the system.

If the system is full one muast retire. The functions
for simulating a retirement have been implemented and tested

* write functions to simulate competence model

More specifically to simulate the hiring and
promoting processes. Tests for both have been
included

* rename functions provoke -> make

* minor tweaks

* fix retirement

We were retiring the person with the maximum retirement date instead of the
min. The last_retirement is renamed to last_retirement_date

* add travis file

* Fix base model (#33)

* write test for promotion that fails

Currently the promotion rate of an individual at level i is based
on the homophily effect of the people at level i. This however means
that the type of the top individual will never have an effect in any
promotion.

So it has been agreed that the promotion of an individual at level i
is affected by the types of individuals at level i + 1.

This test fails with the current implementation

* pass test and tweak mathematical formulation

Closes #32

* update travis.yml

* sort imports