Simulation competence model (#35)
* write & test types in state function

* write functions to simulate retirement

For the competence model we will no longer
calculate the possible states and then getting
the transitios probabilities. Instead we will simulate
the system.

If the system is full one muast retire. The functions
for simulating a retirement have been implemented and tested

* restructure code base

Restructures simulation source code and fixes imports

* write test for competence simulation

* write code for competence simulation and pass test

* fix initial state and add test with old failure

* implement measures

* examples notebooks

* rename last retirement to last retirement date

* simplify promotion, competence is multiplied by a constant

* fix tests after making gamma a constant

* sort imports